Adaptive skjpeg_source_mgr buffer size
skjpeg_source_mgr is implemented as libjpeg-turbo source manager
for SkJpegCodec.  It fixes the buffer size as 1KB size, this is
not efficient for SkStream (either SkFILEStream or SkMemStream)
to read data. The fed data stream has to be long enough, in order
for libjpeg-turbo huffman decoding to by-pass costly end of
bitstream check.

This patch changes this fixed size to adaptive size according to
stream length but limits the size between min 1KB max 512KB.

Antutu 6.1 IMG Decode score is seen to improve 5-10%.

Test: manual test
Change-Id: I7d82503946d42bf75ead124d2848b3533da03ffd
Signed-off-by: cjbao <cathy.bao@intel.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>